requirements.txt
google-cloud-bigquery
streamlit
numpy
plotly
pandas
pandas-gbq